[file] Map data에 있는 model type들 domain내 Entity로 위치 변경
- FindTownUI에 있던 MapDetailCategoryView, MapDetailComponentView Map내 subviews로 변경
- MapViewController에 있었던 MapDetailCategory 관련 로직 View내부로 변경
- AddressSheet 폴더 생성
