+++ FindTown/FindTown/Domain/Entities/Category.swift
@@ -1,0 +1,14 @@
+//
+//  Category.swift
+//  FindTown
+//
+//  Created by 이호영 on 2023/01/07.
+//
+
+import UIKit
+
+struct Category {
+    let image: UIImage
+    let title: String
+    let detailCategories: [DetailCategory]
+}
