+++ FindTown/FindTown/Domain/Entities/DetailCategory.swift
@@ -1,0 +1,14 @@
+//
+//  DetailCategory.swift
+//  FindTown
+//
+//  Created by 이호영 on 2023/01/07.
+//
+
+import UIKit
+
+struct DetailCategory {
+    let color: UIColor
+    let detailTitle: String
+    /// let points
+}
