+++ FindTown/FindTown/Domain/Entities/MapCategory.swift
@@ -1,0 +1,14 @@
+//
+//  MapData.swift
+//  FindTown
+//
+//  Created by 장선영 on 2023/01/05.
+//
+
+import UIKit
+
+/// 임시로 설계해놓은 Map Test Data
+enum MapCategory {
+    case infra
+    case theme
+}
