+++ FindTown/FindTown/Domain/Entities/MapData.swift
@@ -1,0 +1,12 @@
+//
+//  MapCategory.swift
+//  FindTown
+//
+//  Created by 이호영 on 2023/01/07.
+//
+
+import Foundation
+
+struct MapData {
+    let categories: [Category]
+}
